# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FieldForce Live — Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /day-plan:
    get:
      operationId: getUserDayPlan
      tags: [dayplan]
      summary: Get (or create draft for) a user's day plan and ordered visit stops for a date
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: integer }
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: Day plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayPlanDetail"
  /users/{id}/visit-stops:
    post:
      operationId: createVisitStop
      tags: [dayplan]
      summary: Add a visit stop to a user's day plan (resolves address/pin/lat-lng to coordinates)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitStopInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitStop"
  /visit-stops/{stopId}:
    patch:
      operationId: updateVisitStop
      tags: [dayplan]
      summary: Update a visit stop's priority, label, customer code, or status
      parameters:
        - name: stopId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitStopUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitStop"
        "404":
          description: Not found
    delete:
      operationId: deleteVisitStop
      tags: [dayplan]
      summary: Remove a visit stop from a day plan
      parameters:
        - name: stopId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /day-plans/{dayPlanId}/plan-route:
    post:
      operationId: planRoute
      tags: [dayplan]
      summary: >
        Compute priority-driven route order (P1 -> P2 -> P3 bands, shortest-path
        optimized within each band) and persist sequence + totals
      parameters:
        - name: dayPlanId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Planned day plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayPlanDetail"
  /day-plans/{dayPlanId}/publish:
    post:
      operationId: publishDayPlan
      tags: [dayplan]
      summary: Publish (push) the finalized route to the user's device
      parameters:
        - name: dayPlanId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Published day plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayPlanDetail"
  /visit-stops/{stopId}/track-link:
    post:
      operationId: createTrackLink
      tags: [public]
      summary: Create (or reuse) a public customer tracking link for a stop
      parameters:
        - name: stopId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Track link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTrackLink"
  /visit-stops/{stopId}/track-link/revoke:
    post:
      operationId: revokeTrackLink
      tags: [public]
      summary: Revoke the public tracking link for a stop
      parameters:
        - name: stopId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Revoked

  # ---------- Public customer tracking ----------
...
  /dispositions:
    get:
      operationId: listDispositions
      tags: [dispositions]
      summary: List all dispositions for the admin's customer
      responses:
        "200":
          description: Dispositions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Disposition"
    post:
      operationId: createDisposition
      tags: [dispositions]
      summary: Create a disposition label
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispositionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disposition"
  /dispositions/{id}:
    patch:
      operationId: updateDisposition
      tags: [dispositions]
      summary: Update a disposition
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispositionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disposition"
        "404":
          description: Not found
    delete:
      operationId: deleteDisposition
      tags: [dispositions]
      summary: Delete a disposition
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ---------- Config (mobile-facing, no auth) ----------
...
components:
  schemas:
...
    VisitStopPriority:
      type: string
      enum: [P1, P2, P3]
    VisitStopStatus:
      type: string
      enum: [PENDING, EN_ROUTE, REACHED, COMPLETED, SKIPPED]
    VisitStopInputType:
      type: string
      enum: [ADDRESS, PIN, LATLNG]
    DayPlanStatus:
      type: string
      enum: [DRAFT, PUBLISHED]
...
    VisitStop:
      type: object
      required:
        [
          id,
          userId,
          dayPlanId,
          visitDate,
          sequence,
          priority,
          customerCode,
          label,
          inputType,
          rawInput,
          latitude,
          longitude,
          status,
        ]
      properties:
        id: { type: integer }
        userId: { type: integer }
        dayPlanId: { type: integer }
        visitDate: { type: string, format: date }
        sequence: { type: ["integer", "null"] }
        priority:
          $ref: "#/components/schemas/VisitStopPriority"
        customerCode: { type: string }
        label: { type: ["string", "null"] }
        inputType:
          $ref: "#/components/schemas/VisitStopInputType"
        rawInput: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        status:
          $ref: "#/components/schemas/VisitStopStatus"
        plannedArrivalAt: { type: ["string", "null"], format: date-time }
        actualArrivalAt: { type: ["string", "null"], format: date-time }
        contactName: { type: ["string", "null"] }
        contactPhone: { type: ["string", "null"] }
        reachedAt: { type: ["string", "null"], format: date-time }
        startedAt: { type: ["string", "null"], format: date-time }
        closedAt: { type: ["string", "null"], format: date-time }
        dispositionId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
    VisitStopInput:
      type: object
      required: [visitDate, priority, customerCode, inputType, rawInput]
      properties:
        visitDate: { type: string, format: date }
        priority:
          $ref: "#/components/schemas/VisitStopPriority"
        customerCode: { type: string, minLength: 1 }
        label: { type: string }
        inputType:
          $ref: "#/components/schemas/VisitStopInputType"
        rawInput: { type: string, minLength: 1 }
        contactName: { type: string }
        contactPhone: { type: string }
    VisitStopUpdate:
      type: object
      properties:
        priority:
          $ref: "#/components/schemas/VisitStopPriority"
        customerCode: { type: string, minLength: 1 }
        label: { type: string }
        status:
          $ref: "#/components/schemas/VisitStopStatus"
    DayPlanDetail:
      type: object
      required:
        [
          id,
          userId,
          visitDate,
          status,
          totalDistanceMeters,
          totalEtaSeconds,
          stops,
        ]
      properties:
        id: { type: integer }
        userId: { type: integer }
        visitDate: { type: string, format: date }
        status:
          $ref: "#/components/schemas/DayPlanStatus"
        publishedAt: { type: ["string", "null"], format: date-time }
        totalDistanceMeters: { type: number }
        totalEtaSeconds: { type: number }
        stops:
          type: array
          items:
            $ref: "#/components/schemas/VisitStop"
    PublicTrackLink:
      type: object
      required: [token, url, expiresAt]
      properties:
        token: { type: string }
        url: { type: string }
        expiresAt: { type: string, format: date-time }
        revokedAt: { type: ["string", "null"], format: date-time }
...
    Disposition:
      type: object
      required: [id, customerId, label, active, sortOrder]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        label: { type: string }
        active: { type: boolean }
        sortOrder: { type: integer }
    DispositionInput:
      type: object
      required: [label]
      properties:
        label: { type: string, minLength: 1 }
        active: { type: boolean }
        sortOrder: { type: integer }
    DispositionUpdate:
      type: object
      properties:
        label: { type: string, minLength: 1 }
        active: { type: boolean }
        sortOrder: { type: integer }
...